refactor: refactored user score architecture
