+++ src/modules/Main.seeder.ts
@@ -1,9 +1,9 @@
 import { DataSource } from 'typeorm';
 import { Seeder, SeederFactoryManager, runSeeder } from 'typeorm-extension';
-import CourseSeeder from './course/seeders/course.seeder';
-import ActivitySeeder from './activity/seeders/activity.seeder';
+import CourseSeeder from './course/infra/db/seeders/course.seeder';
+import ActivitySeeder from './activity/infra/db/seeders/activity.seeder';
 import UserSeeder from './user/infra/db/seeders/user.seeder';
-import UserScoreSeeder from './user-score/seeders/user-score.seeder';
+import UserScoreSeeder from './user-score/infra/db/seeders/user-score.seeder';
 
 export class MainSeeder implements Seeder {
   async run(
