+++ src/modules/user-activities-answered/services/user-activities-answered.service.spec.ts
@@ -5,8 +5,8 @@ import { DatabaseModule } from '../../../database/database.module';
 import { UserModule } from '../../user/infra/modules/user.module';
 import { User } from '../../user/entity/user.entity';
 import { UserScore } from '../../user-score/entity/user-score.entity';
-import { UserScoreService } from '../../user-score/services/user-score.service';
-import { UserScoreRepository } from '../../user-score/repository/user-score.repository';
+import { UserScoreService } from '../../user-score/infra/services/user-score.service';
+import { UserScoreRepository } from '../../user-score/repositories/user-score.repository';
 import { UserRepository } from '../../user/repositories/user.repository';
 import { JWTProvider } from '../../user/providers/jwt.provider';
 import { UserNotFoundException } from '../../user/domain/dtos/errors/UserNotFound.exception';
