+++ src/modules/user-activities-answered/services/user-activities-answered.service.ts
@@ -7,7 +7,7 @@ import { WrongAnswerException } from '../domain/errors/WrongAnswer.exception';
 import { ActivityNotFoundException } from '../domain/errors/ActivityNotFound.exception';
 import { ActivityAlreadyAnsweredException } from '../domain/errors/ActivityAlreadyAnswered.exception';
 import { UserCourseConcludedRepository } from '../../user-courses-concluded/repository/user-courses-concluded.repository';
-import { UserScoreService } from '../../user-score/services/user-score.service';
+import { UserScoreService } from '../../user-score/infra/services/user-score.service';
 import { CourseRepository } from '../../course/repositories/course.repository';
 
 @Injectable()
