+++ src/modules/user-activities-answered/user-activities-answered.module.ts
@@ -16,8 +16,8 @@ import { UserActivitiesAnsweredController } from './controller/user-activities-a
 import { AuthenticationMiddleware } from '../user/infra/http/middlewares/auth.middleware';
 import { JWTProvider } from '../user/providers/jwt.provider';
 import { UserCourseConcludedRepository } from '../user-courses-concluded/repository/user-courses-concluded.repository';
-import { UserScoreRepository } from '../user-score/repository/user-score.repository';
-import { UserScoreService } from '../user-score/services/user-score.service';
+import { UserScoreRepository } from '../user-score/repositories/user-score.repository';
+import { UserScoreService } from '../user-score/infra/services/user-score.service';
 import { CourseRepository } from '../course/repositories/course.repository';
 import { UserService } from '../user/infra/services/user.service';
 import { HashProvider } from '../user/providers/hash.provider';
