+++ src/modules/user-score/domain/dtos/repositories/user-score.repository.ts
@@ -1,0 +1,7 @@
+import { UserScore } from "../../../infra/db/entity/user-score.entity";
+
+export interface UserScoreRepositoryInterface {
+    findByUserId(user_id: number): Promise<UserScore | null>;
+
+    addScore(user_id: number, score_to_add: number): Promise<void>;
+}
