--- src/modules/user-score/dto/user-score.dto.spec.ts
@@ -1,7 +1,0 @@
-import { UserScoreDto } from './user-score.dto';
-
-describe('UserScoreDto', () => {
-  it('should be defined', () => {
-    expect(new UserScoreDto()).toBeDefined();
-  });
-});
