--- src/modules/user-score/dto/user-score.dto.ts
@@ -1,1 +1,0 @@
-export class UserScoreDto {}
