+++ src/modules/user-score/infra/db/entity/user-score.entity.spec.ts
