+++ src/modules/user-score/infra/db/entity/user-score.entity.ts
@@ -1,4 +1,4 @@
-import { User } from '../../../modules/user/entity/user.entity';
+import { User } from "../../../../user/infra/db/entity/user.entity";
 import {
   Column,
   CreateDateColumn,
