+++ src/modules/user-score/infra/db/repositories/user-score.repository.spec.ts
