+++ src/modules/user-score/infra/db/seeders/user-score.seeder.ts
