+++ src/modules/user-score/infra/http/controllers/user-score.controller.spec.ts
