+++ src/modules/user-score/infra/modules/user-score.module.ts
@@ -4,20 +4,20 @@ import {
   NestModule,
   RequestMethod,
 } from '@nestjs/common';
-import { UserScoreService } from './services/user-score.service';
+import { UserScoreService } from '../services/user-score.service';
 import {
   ConjunctUserScoreController,
   IndividualUserScoreController,
-} from './controller/user-score.controller';
-import { User } from '../user/entity/user.entity';
+} from '../http/controllers/user-score.controller';
+import { User } from '../../../user/infra/db/entity/user.entity';
 import { DatabaseModule } from 'src/database/database.module';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { UserModule } from '../user/infra/modules/user.module';
-import { UserScoreRepository } from './repository/user-score.repository';
-import { UserScore } from './entity/user-score.entity';
-import { UserRepository } from '../user/repositories/user.repository';
-import { AuthenticationMiddleware } from '../user/infra/http/middlewares/auth.middleware';
-import { JWTProvider } from '../user/providers/jwt.provider';
+import { UserModule } from '../../../user/infra/modules/user.module';
+import { UserScoreRepository } from '../db/repositories/user-score.repository';
+import { UserScore } from '../db/entity/user-score.entity';
+import { UserRepository } from '../../../user/infra/db/repositories/user.repository';
+import { AuthenticationMiddleware } from '../../../user/infra/http/middlewares/auth.middleware';
+import { JWTProvider } from '../../../user/infra/providers/jwt.provider';
 
 @Module({
   imports: [
