+++ src/modules/user-score/infra/services/user-score.service.ts
@@ -1,8 +1,8 @@
 import { Injectable } from '@nestjs/common';
-import { UserScoreRepository } from '../repository/user-score.repository';
-import { UnprocessableDataException } from '../../../shared/domain/errors/UnprocessableData.exception';
-import { UserNotFoundException } from '../../user/domain/dtos/errors/UserNotFound.exception';
-import { FindTopScoresResponseDTO } from '../../user/domain/dtos/requests/FindTopScores.request.dto';
+import { UserScoreRepository } from '../repositories/user-score.repository';
+import { UnprocessableDataException } from '../../../../shared/domain/errors/UnprocessableData.exception';
+import { UserNotFoundException } from '../../../user/domain/dtos/errors/UserNotFound.exception';
+import { FindTopScoresResponseDTO } from '../../../user/domain/dtos/requests/FindTopScores.request.dto';
 import { UserRepository } from '../../user/repositories/user.repository';
 
 @Injectable()
